api/requirements.txt
astroid==2.14.1
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
dill==0.3.6
dnspython==2.3.0
Flask==2.2.2
Flask-Cors==3.0.10
idna==3.4
importlib-metadata==6.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pycodestyle==2.10.0
pylint==2.16.1
pymongo==4.3.3
python-dotenv==0.21.1
requests==2.28.2
six==1.16.0
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.13.0
